+++ server-worker/src/main/java/com/aliyun/emr/rss/service/deploy/worker/FileWriter.java
@@ -25,7 +25,9 @@ import java.util.List;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.concurrent.atomic.AtomicInteger;
+import java.util.concurrent.atomic.AtomicReference;
 
+import com.aliyun.emr.rss.service.deploy.worker.storage.*;
 import io.netty.buffer.ByteBuf;
 import io.netty.buffer.CompositeByteBuf;
 import org.slf4j.Logger;
@@ -40,9 +42,6 @@ import com.aliyun.emr.rss.common.network.server.MemoryTracker;
 import com.aliyun.emr.rss.common.protocol.PartitionSplitMode;
 import com.aliyun.emr.rss.common.protocol.PartitionType;
 import com.aliyun.emr.rss.common.protocol.StorageInfo;
-import com.aliyun.emr.rss.service.deploy.worker.storage.*;
-import com.aliyun.emr.rss.service.deploy.worker.storage.local.LocalFlushTask;
-import com.aliyun.emr.rss.service.deploy.worker.storage.local.LocalFlusher;
 
 /*
  * Note: Once FlushNotifier.exception is set, the whole file is not available.
@@ -87,6 +86,26 @@ public final class FileWriter implements DeviceObserver {
               + mountPoint + " : " + diskStatus));
     }
     deviceMonitor.unregisterFileWriter(this);
+  }
+
+  static class FlushNotifier {
+    final AtomicInteger numPendingFlushes = new AtomicInteger();
+    final AtomicReference<IOException> exception = new AtomicReference<>();
+
+    void setException(IOException e) {
+      exception.set(e);
+    }
+
+    boolean hasException() {
+      return exception.get() != null;
+    }
+
+    void checkException() throws IOException {
+      IOException e = exception.get();
+      if (e != null) {
+        throw e;
+      }
+    }
   }
 
   private final FlushNotifier notifier = new FlushNotifier();
