--- server-worker/src/main/java/com/aliyun/emr/rss/service/deploy/worker/FlushNotifier.java
@@ -1,42 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.aliyun.emr.rss.service.deploy.worker;
-
-import java.io.IOException;
-import java.util.concurrent.atomic.AtomicInteger;
-import java.util.concurrent.atomic.AtomicReference;
-
-public class FlushNotifier {
-  public final AtomicInteger numPendingFlushes = new AtomicInteger();
-  public final AtomicReference<IOException> exception = new AtomicReference<>();
-
-  public void setException(IOException e) {
-    exception.set(e);
-  }
-
-  public boolean hasException() {
-    return exception.get() != null;
-  }
-
-  public void checkException() throws IOException {
-    IOException e = exception.get();
-    if (e != null) {
-      throw e;
-    }
-  }
-}
