+++ server-worker/src/test/java/com/aliyun/emr/rss/service/deploy/worker/FileWriterSuiteJ.java
@@ -70,7 +70,6 @@ import com.aliyun.emr.rss.common.protocol.PartitionType;
 import com.aliyun.emr.rss.common.protocol.StorageInfo;
 import com.aliyun.emr.rss.common.util.ThreadUtils;
 import com.aliyun.emr.rss.common.util.Utils;
-import com.aliyun.emr.rss.service.deploy.worker.storage.DeviceMonitor$;
 import com.aliyun.emr.rss.service.deploy.worker.storage.local.LocalFlusher;
 
 import static org.junit.Assert.assertEquals;
